+++ packages/lodash-pro/src/isArray/index.ts
@@ -1,5 +1,0 @@
-import { EnumDataType, is } from '../is'
-
-export function isArray(data: unknown): data is Array<any> {
-  return is(data, EnumDataType.array)
-}