+++ packages/lodash-pro/src/isBoolean/index.ts
@@ -1,5 +1,0 @@
-import { EnumDataType, is } from '../is'
-
-export function isBoolean(data: unknown): data is boolean {
-  return is(data, EnumDataType.boolean)
-}