+++ packages/lodash-pro/src/isFunction/index.ts
@@ -1,5 +1,0 @@
-import { EnumDataType, is } from '../is'
-
-export function isFunction(data: unknown): data is Function {
-  return is(data, EnumDataType.function)
-}