+++ packages/lodash-pro/src/isNil/index.ts
@@ -1,6 +1,0 @@
-import { isNull } from '../isNull'
-import { isUndefined } from '../isUndefined'
-
-export function isNil(val: unknown): val is null | undefined {
-  return isNull(val) || isUndefined(val)
-}