+++ packages/lodash-pro/src/isNull/index.ts
@@ -1,5 +1,0 @@
-import { EnumDataType, is } from '../is'
-
-export function isNull(data: unknown): data is null {
-  return is(data, EnumDataType.null)
-}