+++ packages/lodash-pro/src/isNumber/index.ts
@@ -1,5 +1,0 @@
-import { EnumDataType, is } from '../is'
-
-export function isNumber(data: unknown): data is number {
-  return is(data, EnumDataType.number)
-}