+++ packages/lodash-pro/src/isObject/index.ts
@@ -1,5 +1,0 @@
-import { EnumDataType, is } from '../is'
-
-export function isObject(data: unknown): data is Record<any, any> {
-  return is(data, EnumDataType.object)
-}