+++ packages/lodash-pro/src/isString/index.ts
@@ -1,5 +1,0 @@
-import { EnumDataType, is } from '../is'
-
-export function isString(data: unknown): data is string {
-  return is(data, EnumDataType.string)
-}