+++ packages/lodash-pro/src/isUndefined/index.ts
@@ -1,5 +1,0 @@
-import { EnumDataType, is } from '../is'
-
-export function isUndefined(data: unknown): data is undefined {
-  return is(data, EnumDataType.undefined)
-}