+++ packages/flexible-pro/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @minko-fe/flexible-pro
 
+## 0.1.21
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @minko-fe/lodash-pro@0.2.3
+
 ## 0.1.20
 
 ### Patch Changes