+++ packages/postcss-config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @minko-fe/postcss-config
 
+## 0.1.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @minko-fe/lodash-pro@0.2.3
+
 ## 0.1.21
 
 ### Patch Changes